case insensitive search

+++ compare-healthcare-api/Models/Country.cs
@@ -8,7 +8,7 @@ public class Country
 	//https://stackoverflow.com/questions/450238/to-underscore-or-to-not-to-underscore-that-is-the-question
   public string countryName { get; set; }
   public int rank { get; set; }
-  public int monthWaitingListDelay { get; set; }
+  public float monthWaitingListDelay { get; set; }
   public float aAndEHoursWait { get; set; }
   public CostOfTreatment costOfTreatment { get; set; }
   public string customerOpinion { get; set; }
